Micha created from utils branch merged with envelope branch

+++ envelope.h
@@ -11,10 +11,11 @@
 #define envelope_h
 
 #include "m_pd.h"
+#include "grain.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
-#include "grain.h"
+
 
 #ifdef __cplusplus
 extern "C" {
@@ -59,7 +60,8 @@ typedef struct window
     t_sample *window_samples_table;
 }window;
 
-envelope *envelope_new(int attack, int decay, int sustain, int key_pressed, int release);
+envelope *envelope_new(int attack, int decay, float sustain, int key_pressed, int release);
+
 float gauss(grain x, int sample);
 void envelope_free(envelope *x);
